test(rtc_data_service): refactor: move tests into ecalls and web_api modules
These are mainly renames, except for:

* Move the `local_attestation` test out of `exec_token` and into
`ecalls` instead

* Introduce a top-level `main.rs` module, to allow sharing test helpers
between `ecalls` and `web_api`
