+++ rtc_data_service/tests/ecalls/mod.rs
@@ -1,0 +1,3 @@
+//! ECALL tests
+
+mod local_attestation;
