+++ rtc_data_service/tests/main.rs
@@ -1,0 +1,4 @@
+//! Top-level test module
+
+mod ecalls;
+mod web_api;
