+++ rtc_data_service/tests/web_api/exec_token.rs
@@ -9,8 +9,6 @@ use actix_web::{http, test};
 
 use rtc_data_service::data_enclave_actor::DataEnclaveActor;
 use rtc_data_service::exec_token;
-use rtc_uenclave::EnclaveConfig;
-use sgx_types::sgx_status_t;
 
 #[actix_rt::test]
 async fn data_service_exec_token_ok() {
@@ -66,26 +64,4 @@ async fn data_service_exec_token_ok() {
         nonce: vec![7; 24],
     };
     assert_eq!(expected, actual)
-}
-
-#[test]
-fn test_local_attestation_success() {
-    let auth_enclave = rtc_uenclave::RtcAuthEnclave::init(EnclaveConfig {
-        lib_path: "/root/rtc-data/rtc_auth_enclave/build/bin/enclave.signed.so".to_string(),
-        ..Default::default()
-    })
-    .unwrap();
-
-    let data_enclave = rtc_uenclave::RtcDataEnclave::init(EnclaveConfig {
-        lib_path: "/root/rtc-data/rtc_data_enclave/build/bin/enclave.signed.so".to_string(),
-        ..Default::default()
-    })
-    .unwrap();
-
-    let res = data_enclave.local_attestation(auth_enclave.geteid());
-    assert_eq!(res, sgx_status_t::SGX_SUCCESS);
-
-    // TODO: Integration test for message sending
-    // We should consider moving the integration tests for enclave interaction into rtc_uenclave
-    // since these tests does not need anything from the data_service
 }
