+++ rtc_data_service/tests/web_api/mod.rs
@@ -1,0 +1,5 @@
+//! Web API tests
+
+mod attestation;
+mod data_upload;
+mod exec_token;
